grsec: Apply grsecurity-3.1-3.14.40-201504302118.patch
commit 1b98f4762286dbe86e9f0af00d559a79530c355c
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Apr 30 19:40:23 2015 -0400

    Fix a refcount underflow reported by Jan Hejl caused by GRKERNSEC_CHROOT_RENAME
    We failed to increment the tree in copy_fs_struct, so users of clone with
    CLONE_FS who then exited caused a decrement of the refcounts with no associated
    increment.  This would generally occur on / and took at least a month on
    a heavily-loaded system to trigger.  It shouldn't cause any security problems
    as no freeing is associated with the refcount, and the only interesting value
    is 0, which permits renames in that tree.  I've tested this fixed implementation
    to ensure the refcounts are under control in both directions.

 fs/fs_struct.c |    1 +
 1 files changed, 1 insertions(+), 0 deletions(-)

commit b46f1bc40a99043794c760fafea335fe1ca087fb
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Apr 30 08:39:48 2015 -0400

    compile fix

 arch/x86/kernel/apic/io_apic.c   |    2 +-
 drivers/xen/events/events_base.c |    6 +++---
 2 files changed, 4 insertions(+), 4 deletions(-)

commit e8faefbe3f7eb3f2e98987b95d292d2e407402a1
Merge: 6e09ec7 c926430
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Apr 30 08:38:40 2015 -0400

    Merge branch 'pax-stable2' into grsec-stable2

commit c926430b0d52397af37a63b15ddb15d9ece60a46
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Apr 30 08:38:17 2015 -0400

    Update to pax-linux-3.14.40-test44.patch:
    - fixed compilation errors due to some overzealous constification of irq_chip variables, reported by spender

 arch/x86/kernel/apic/io_apic.c   |    4 ++--
 drivers/xen/events/events_base.c |    6 +++---
 2 files changed, 5 insertions(+), 5 deletions(-)

commit 6e09ec7529a9801166a182648a1c808755bd4199
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Apr 30 07:18:06 2015 -0400

    compile fix

 drivers/xen/events/events_base.c |    6 +++---
 1 files changed, 3 insertions(+), 3 deletions(-)

commit bca9c824b0e8ae0c7794e04307c4af9247905dcc
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Apr 30 07:14:18 2015 -0400

    Update size_overflow hash

 .../size_overflow_plugin/size_overflow_hash.data   |    1 +
 1 files changed, 1 insertions(+), 0 deletions(-)

commit 0364548df8847766844a2d91d397d09f9991f4b8
Author: Brad Spengler <spender@grsecurity.net>
Date:   Wed Apr 29 22:34:18 2015 -0400

    compile fix

 arch/x86/kernel/apic/io_apic.c |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)

commit d2e5fa2378b8ff37963278120a3519cbcf290178
Merge: b895f5b e727cfa
Author: Brad Spengler <spender@grsecurity.net>
Date:   Wed Apr 29 21:53:16 2015 -0400

    Merge branch 'pax-stable2' into grsec-stable2

    Conflicts:
    	tools/gcc/gcc-common.h

commit e727cfaee885f426fc3b760eed34a83c894114f4
Author: Brad Spengler <spender@grsecurity.net>
Date:   Wed Apr 29 21:50:40 2015 -0400

    Update to pax-linux-3.14.40-test43.patch:
    - ported all plugins to gcc-5 except size overflow
    - fixed integer signedness mixup in mmc error code handling, caught by the size overflow plugin, reported by Tom Seewald
    - fixed a few section conflicts exposed by LTO
    - fixed the latent entropy plugin for LTO
    - fixed the stackleak plugin for LTO

 arch/x86/include/asm/page_64.h            |    4 +-
 arch/x86/include/asm/uaccess.h            |    8 +-
 arch/x86/kernel/apic/io_apic.c            |    4 +-
 arch/x86/kernel/apm_32.c                  |    2 +-
 arch/x86/kernel/reboot.c                  |    2 +-
 arch/x86/mm/pgtable.c                     |    2 +-
 crypto/zlib.c                             |    4 +-
 drivers/acpi/bus.c                        |    4 +-
 drivers/acpi/ec.c                         |    2 +-
 drivers/acpi/pci_slot.c                   |    2 +-
 drivers/acpi/processor_core.c             |    2 +-
 drivers/acpi/processor_driver.c           |    2 +-
 drivers/acpi/sleep.c                      |    2 +-
 drivers/acpi/thermal.c                    |    2 +-
 drivers/acpi/video.c                      |    2 +-
 drivers/char/i8k.c                        |    2 +-
 drivers/char/sonypi.c                     |    2 +-
 drivers/firmware/efi/runtime-map.c        |    2 +-
 drivers/firmware/google/gsmi.c            |    2 +-
 drivers/firmware/google/memconsole.c      |    2 +-
 drivers/firmware/memmap.c                 |    2 +-
 drivers/hwmon/acpi_power_meter.c          |    2 +-
 drivers/input/touchscreen/htcpen.c        |    2 +-
 drivers/md/raid5.c                        |   12 +-
 drivers/mfd/kempld-core.c                 |    2 +-
 drivers/net/ethernet/via/via-rhine.c      |    2 +-
 drivers/pci/pcie/portdrv_pci.c            |    2 +-
 drivers/platform/chrome/chromeos_pstore.c |    2 +-
 drivers/platform/x86/compal-laptop.c      |    2 +-
 drivers/platform/x86/hdaps.c              |    2 +-
 drivers/platform/x86/ibm_rtl.c            |    2 +-
 drivers/platform/x86/intel_oaktrail.c     |    2 +-
 drivers/platform/x86/msi-laptop.c         |    2 +-
 drivers/platform/x86/samsung-laptop.c     |    2 +-
 drivers/platform/x86/samsung-q10.c        |    2 +-
 drivers/platform/x86/sony-laptop.c        |    2 +-
 drivers/pnp/pnpbios/core.c                |    2 +-
 drivers/thermal/x86_pkg_temp_thermal.c    |    2 +-
 drivers/xen/events/events_base.c          |    6 +-
 include/linux/compiler-gcc5.h             |    2 -
 include/linux/compiler.h                  |    2 +
 include/linux/mmc/core.h                  |    2 +-
 include/linux/syscalls.h                  |    5 +-
 mm/madvise.c                              |    4 +-
 net/l2tp/l2tp_ip.c                        |    2 +-
 net/l2tp/l2tp_ip6.c                       |    2 +-
 net/netfilter/nft_compat.c                |    9 +--
 security/Kconfig                          |    2 +-
 tools/gcc/Makefile                        |    2 +-
 tools/gcc/colorize_plugin.c               |    4 +-
 tools/gcc/constify_plugin.c               |   11 ++-
 tools/gcc/gcc-common.h                    |  166 +++++++++++++++++++++++++++--
 tools/gcc/kallocstat_plugin.c             |   11 ++-
 tools/gcc/kernexec_plugin.c               |   63 ++++++++----
 tools/gcc/latent_entropy_plugin.c         |   21 +++--
 tools/gcc/stackleak_plugin.c              |   52 ++++++++--
 tools/gcc/structleak_plugin.c             |   25 ++++-
 57 files changed, 361 insertions(+), 130 deletions(-)

Signature-tree: 68b4e50520f3b35c92efbfb1accce8bba8d581d1

+++ drivers/md/raid5.c
@@ -942,23 +942,23 @@ async_copy_data(int frombio, struct bio *bio, struct page *page,
 	struct bio_vec bvl;
 	struct bvec_iter iter;
 	struct page *bio_page;
-	int page_offset;
+	s64 page_offset;
 	struct async_submit_ctl submit;
 	enum async_tx_flags flags = 0;
 
 	if (bio->bi_iter.bi_sector >= sector)
-		page_offset = (signed)(bio->bi_iter.bi_sector - sector) * 512;
+		page_offset = (s64)(bio->bi_iter.bi_sector - sector) * 512;
 	else
-		page_offset = (signed)(sector - bio->bi_iter.bi_sector) * -512;
+		page_offset = (s64)(sector - bio->bi_iter.bi_sector) * -512;
 
 	if (frombio)
 		flags |= ASYNC_TX_FENCE;
 	init_async_submit(&submit, flags, tx, NULL, NULL, NULL);
 
 	bio_for_each_segment(bvl, bio, iter) {
-		int len = bvl.bv_len;
-		int clen;
-		int b_offset = 0;
+		s64 len = bvl.bv_len;
+		s64 clen;
+		s64 b_offset = 0;
 
 		if (page_offset < 0) {
 			b_offset = -page_offset;
